+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Versions with only mechanical changes will be omitted from the following list.
 * Add support for microseconds timestamps serde serialization for `NaiveDateTime`.
 * Add support for optional timestamps serde serialization for `NaiveDateTime`.
 * Fix build for wasm32-unknown-emscripten (@yu-re-ka #593)
+* Make `ParseErrorKind` public and available through `ParseError::kind()` (#588)
 * Implement `DoubleEndedIterator` for `NaiveDateDaysIterator` and `NaiveDateWeeksIterator`
 * Fix panicking when parsing a `DateTime` (@botahamec)